Adjust update to P[rain/snow] to account for all var types
Previous syntax caused issues in cases where T was scalar/time series rather than grid. Yes, this allocates to new memory rather than overwriting the previous value in the same memory space (this should be improved later), but for user friendliness out-of-the-box, I think this is the best option for now.